Combine project and referral validation into validationUtil.ts
Related to #1222
